--- src/main/kotlin/Calculator/CalculatorException.kt
@@ -1,3 +1,0 @@
-package Calculator
-
-class CalculatorException(message: String) : Exception()
