+++ src/main/kotlin/step1/CalculatorException.kt
@@ -1,3 +1,3 @@
-package calculator
+package step1
 
 class CalculatorException(message: String) : IllegalArgumentException(message)
