+++ src/test/kotlin/step1/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package calculator
+package step1
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
